+++ src/tests/functional/shared_test_classes/src/base/utils/compare_results.cpp
@@ -3,8 +3,8 @@
 //
 
 #include "ngraph/ops.hpp"
-#include "ngraph_ops/augru_cell.hpp"
-#include "ngraph_ops/augru_sequence.hpp"
+#include "ov_ops/augru_cell.hpp"
+#include "ov_ops/augru_sequence.hpp"
 
 #include "shared_test_classes/base/layer_test_utils.hpp"
 #include "shared_test_classes/base/utils/compare_results.hpp"
@@ -81,7 +81,7 @@ CompareMap getCompareMap() {
 #include "openvino/opsets/opset9_tbl.hpp"
 #include "openvino/opsets/opset10_tbl.hpp"
 
-#include "ngraph_ops/opset_private_tbl.hpp"
+#include "ov_ops/opset_private_tbl.hpp"
 #undef _OPENVINO_OP_REG
     };
     return compareMap;
