+++ src/tests/ngraph_helpers/lpt_ngraph_functions/src/recurrent_cell_function.cpp
@@ -5,7 +5,7 @@
 #include "lpt_ngraph_functions/recurrent_cell_function.hpp"
 
 #include <ngraph/opsets/opset1.hpp>
-#include "ngraph_ops/type_relaxed.hpp"
+#include "ov_ops/type_relaxed.hpp"
 #include "low_precision/network_helper.hpp"
 #include "low_precision/rt_info/precision_preserved_attribute.hpp"
 #include "low_precision/rt_info/intervals_alignment_attribute.hpp"
